wechat-channels-parser/pyproject.toml
[project]
name = "wechat-channels-parser"
version = "0.1.0"
description = "企业微信微信客服 → 视频号链接解析服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pycryptodome>=3.21",
    "xmltodict>=0.14",
    "pydantic-settings>=2.6",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
